requirements.txt
Django==5.2.5
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2

# REST framework & auth
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1

# Admin UI enhancements
django-jazzmin==3.0.1

# Optional admin charts dependencies (included because present in venv)
django-admin-charts==1.5.0
django-bower==5.2.0
django-memoize==2.3.1
django-multiselectfield==1.0.1
django-nvd3==0.10.1
python-nvd3==0.16.0
datetime-truncate==1.1.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
text-unidecode==1.3
six==1.17.0

# Templates
Jinja2==3.1.6
MarkupSafe==3.0.2

# Database (PostgreSQL optional)
psycopg2-binary==2.9.10
